chore(mds): Refactor the master's masterconn
Shadows and Metaloggers use the struct masterconn to represent a
connection to the active Master server in the current architecture.
This commit, partially refactors the masterconn.cc file. The main
changes are:

- Improve readability.
  - Semi camel case variable names.
  - Rename variables to be more consistent with their real usage.
- Add documentation.
- Reduce enums size.
- Remove some clang-tidy warnings (will continue).
- Prepare the struct to be split later into .h and .cc files, for
  better separation between interface and implementation.

The change was limited to only affect masterconn.cc to ease the review.
Next commits will continue improving the related files.

Signed-off-by: guillex <guillex@leil.io>