chore(mds): Move functions to MasterConn
The functions associated with the MasterConn struct are now member
funtions, removing the need to pass the pointer to the object every
time, and making easier to have multiple connections in the future.
The funcions were also renamed to be semi camel case.

The global functions were grouped and moved to the anonymous namespace.

Signed-off-by: guillex <guillex@leil.io>