chore(mds): Remove static string init warnings
Clang-tidy was complaining about some initializations of std::string at
compile time (may throw and can not be catched). This commit removes
the warnings by creating static inline functions returning the correct
values for Metalogger and Shadow.

The change uses static storage and const references to be as efficient
as the previous approach.

Signed-off-by: guillex <guillex@leil.io>