refactor: Split OutputBuffer inner buffer
The main goal of this change is to prepare the OutputBuffer to emplace
bigger pieces of data at once during pread calls. This could not be
achieved previously because the buffer should have a sequence of
header, CRC, data block, header, crc and so on.

This commit also adds a little bit more of readability when putting
data into the OutputBuffer.

Signed-off-by: Dave <dave@leil.io>