refactor: Create OutputBuffer::Buffer class
This commit mainly creates the class OutputBuffer::Buffer. This is a
generic class to implement the three types of buffers an OutputBuffer
has. Taking this into consideration, the implementation of the
OutputBuffer could be made simpler and more readable.

As a side change, the WriteStatus enum modernized to use the enum
class.

Signed-off-by: Dave <dave@leil.io>