test: Improve unit tests for OutputBuffer class
This commit modernizes the existing unit test for the OutputBuffer and
adds other three tests for the same class:
- outputBufferCheckCrcTest: to check CRC calculation from the bare
buffer.
- outputBufferCopyValueTest: to check the copy of a value from the
buffer and the CRC of it afterwards.
- outputBufferSeveralBlocksInBufferTest: to check the expected behavior
of the OutputBuffer when several blocks are added to it, with an
uncomplete last block. Header, CRC and data are checked in this test.

Signed-off-by: Dave <dave@leil.io>