feat(chunkserver): Parallelize reads per csentry (#319)
This commit targets making the ChunkserverEntry capable of sending
several read jobs at a time to the background workers.

The main part of the change was to make the readContinue function
able to send several read jobs into the jobs pool. To make this change
consistent, it was needed to change:
- the way failing read jobs are handled.
- the way the ChunkserverEntry is closed.

The job pool structure needed to be extended to fulfill the new
requirements, as well as the OutputBuffer.

In order to configure the new behavior, the option
MAX_PARALLEL_HDD_READ_JOBS_PER_CS_ENTRY was added to the chunkserver
options. As a side change, the following constants were also
increased:
- kDefaultMaxBackgroundJobsPerNetworkWorker, to make sure the jobs pool
could handle a larger number of parallel requests by default. The
related parameter BGJOBSCNT_PER_NETWORK_WORKER default value was also
changed.
- BuffersPool::kMaxSize to make sure there will be enough room to store
the higher number of buffers used at the same time.

Signed-off-by: Dave <dave@leil.io>

Co-authored-by: Crash <crash@leil.io>
Co-authored-by: guillex <guillex@leil.io>