fix: Fix SQLite issue
An issue appeared when several processes/threads started doing
concurrent operations an SQLite database. The issue originated from
having different ReadRecords for each those processes while also doing
writes on the same file the cache was about. Solution is to clear the
cache after chunk is locked and prepared to be written.

Signed-off-by: Dave <dave@leil.io>