refactor(chunkserver): Split OutputBuffer inner buffer (#328)
The main goal of this change is to prepare the `OutputBuffer` to emplace
bigger pieces of data at once during pread calls. This could not be
achieved previously because the buffer should have a sequence of
header, CRC, data block, header, CRC and so on.

The class `OutputBuffer::Buffer` is created for this purpose. This is a
generic class to implement the three types of buffers an `OutputBuffer`
has. Taking this into consideration, the implementation of the
`OutputBuffer` could be made simpler and more readable.

Side changes:
- increase readability of putting data into the `OutputBuffer`.
- save the allocation of a block of zeros when asking for block beyond
current number of blocks in chunk.
- modernize the `WriteStatus` enum.
- change `BuffersPool` to adapt to the new format of the `OutputBuffer`.
Keep track of the number of block in the buffers instead of the number
of buffers, and log it from time to time.
- add documentation for the `OutputBuffer` class.
- improve the unit tests suite of the `OutputBuffer` class.

Signed-off-by: Dave <dave@leil.io>