test: Build clang in pipeline instead of as a test
The clang test does not use the vcpkg libraries, causing the test to
fail because fmt:println isn't implemented in the system library that is
being used.
Due to complexities in including vcpkg into the test that works both
locally and in Docker, it would be better to do it in a Jenkins
pipeline, which can benefit in caching. It is the first stage, because
it should be relatively quick.

Signed-off-by: Urmas Rist <urmas@leil.io>