feat(all): Unify inode numbers type
The inode numbers type (currently uint32_t) is not easily replaceable,
there are different kinds of definitions across the code, like:
- typedef uint32_t SaunaClient::Inode
- using InodeId = uint32_t; // In the ACLs
- plain uint32_t
- etc.

This commit starts to unify the type for inode numbers across the
complete codebase. The type is now defined in the new file type_defs.h
in a way compatible with C++ and C (for the client lib). All other
references to an inode number should use this type (Inode).

Signed-off-by: guillex <guillex@leil.io>