fix: Fix SQLite issue (#337)
An issue appeared when several processes/threads started doing
concurrent operations on a SQLite database in our system. The issue
originated from having different ReadRecords for each those processes
while also doing writes on the same file the cache was about. Solution
is to clear the cache after chunk is locked and prepared to be written.

A couple of tests were added to test the fix.

Signed-off-by: Dave <dave@leil.io>

Co-authored-by: aNeutrino <pm@leil.io>