feat(mount): Reduce write memory usage (#343)
This PR targets reducing the high memory usage of our client after
some write operation is performed.

The usage of new/delete[] on the write cache blocks during intensive
writes could lead to an increased memory fragmentation. The
substitution with an analogous vector seeks to reduce it, while
reducing the complexity of maintaining the code.

This commit adds the malloctrimperiod options to the FUSE client. This
option targets faster release of the memory used by the client the
Linux kernel. By default, it keeps current behavior (let the OS reclaim
the memory).

Signed-off-by: Dave <dave@leil.io>

Co-authored-by: Rolando Sánchez Ramos <rolysr@leil.io>