fix(chunkserver): Fix long rebalancing/replicate
The current rebalancing and replicate algorithm is based on the master
message to the chunkserver to replicate a specific part of the chunk.

That job is posted by the network workers of the chunkserver listening
to master, but processed by the job workers. At the moment of posting
the job to the JobPool, the buffer to be deserialized contains the
expected list of sources to get the chunks from, but this buffer is
relative to the master connection which may change after several other
incoming packets. So, when a job worker gets the actual job, the data
inside the buffers could have changed.

The proposed fix is to deserialize before posting the job in the JobPool.

A couple of tests were added to check there are no errors during
rebalancing and replication.

Signed-off-by: Dave <dave@leil.io>

Co-authored-by: Crash <crash@leil.io>