fix(master): Large trash causing crash
The current FUSE metadata mount is reading the entire trash from master,
however the protocol is limited to UINT32_MAX. When the trash
data exceeds this, it's causing wrapping to the uint32 variable, which
is then used to determine the malloc size of the memory to write the
data, causing a buffer overflow. Besides, it will allocate a lot of
memory for large trash sizes and hang master as it tries to respond with
the *ENTIRE* trash to client.

The fix to this is limiting the amount of files that master
will return. Currently, this is 1 million. There are some drawbacks with
this approach: To see newer trash entries, you will need to delete or
undelete files first.

A more general solution is actually implemented, which allows buffering
the trash from client (SAU_CLTOMA_FUSE_GETTRASH) using max_entries.
Unfortunately, this was never implemented for the FUSE meta mount.
Another commit should be made to make the clients use this. But the
focus of this commit is fixing CLTOMA_FUSE_GETTRASH packet, so metadata
clients using this will not crash master.

Signed-off-by: Urmas Rist <urmas@leil.io>