fix: Fix test_optimizing_partial_stripes
The test test_optimizing_partial_stripes was failing from time to time
with some noteworthy probability. This test checked the write algorithm
to make sure the client tries to send full stripes every time it writes
to the chunkservers and waits some time before writing a partial stripe
(in the case when nothing else needs to be written).

The failing check was the first one, the test expects the checker
client to perceive the test file size to be stripe size (as the rest
of the write should wait to see it could be filled) but it is directly
getting it as an extra block over the stripe size. This means the
writer client actually wrote the extra block when it was not supposed
to.

The detected failing scenario is the following:
- the first stripe is correctly cached and sent to the chunkservers
as expected, this occurs after the next block starts its caching.
- when the full stripe finishes its write to the chunkservers, two
things happen when that make the test to fail: the next block is fully
in the inode's datachain and the next next block (5th one) has not even
been created and appended to the datachain.
- the 4th block is then moved the ChunkWriter because it is complete,
this moving makes the datachain empty.
- the can_expect_next_block variable then turns false and this in turn
causes the ChunkWriter to start the operation to send this 4th block
right away, which leads to the unexpected file size.

As a retionale, the key error here is that the can_expect_next_block
should be true in this case: we have just remove the last block from
the datachain, and that stripe could still be completed. So the
proposed solution is to patch this specific case, i.e add a method for
the inodedata to check whether it as recently written and we have taken
that last block already. The can_expect_next_block can check then this
case to cover it.

Signed-off-by: Dave <dave@leil.io>