refactor(fuse): Refactor reading root directory from meta mount point
Streaming trash files from master is not as easy to implement on first
thought, due to the complexity of the existing code in the fuse
implementation. Thus, this commit is the first in the series of
refactoring the metadata readdir code from C to C++.

The commit aims to refactor reading the metadata root directory, while
trying not to touch anything else. More commits should follow to further
refactor the metadata mountpoint.

Signed-off-by: Urmas Rist <urmas@leil.io>