fix: Fix replicate related chunkserver locking (#381)
This change targets removing an unprotected access (chunk isn't locked
nor the gChunksMapMutex) to the chunk's data file name that is being
tested by the tester thread in the chunkserver.

The chunks structure from the disks class needs to be protected with
the gTestsMutex mutex. This protection was missing in this function and
is being added by this change.

The max_wait_time had static modifier which apparently tried to make it
shared among all the ChunkReplicator instances created. Those instances
could be created from different threads simultaneously, causing a data
race condition warning while doing an intensive replicate
(rebalance/replicate) workload. The proposal is to leave it as a
regular local const variable.

Merge pull request #381 from leil-io/fix-chunkserver-locking-replicate

Signed-off-by: Dave <dave@leil.io>