fix: Fix client recover when master disconnects (#383)
From commit 88f6a0e, changes were introduced to modify the behavior of
the client when the master disconnects. These changes ensured that the
client would remain unusable but would not crash in such scenarios.

However, the changes caused an issue in the Windows client, preventing
it from recovering from a read/write operation when the master
disconnects. As a result, the client was unable to correctly complete
the operation once the master reconnected.

To address this, the best course of action is to undo those changes,
ensuring that the Windows client maintains a more consistent and
reliable behavior. For now in the general case of master disconnection,
the Windows client will not crash as long as it is not forced to it,
allowing it to work as expected when the maximum number of retries is
reached or the master connects again.

Signed-off-by: rolysr <rolysr@leil.io>