feat: Disable creation of folders for full disk (#346)
Previously, metadata creation was not considering the available space
in the cluster. Even when the space was depleted, it was possible to
create folders.

This commit introduces `CREATE_EMPTY_FOLDERS_WHEN_SPACE_DEPLETED`
option in the master server to disable the creation of metadata for
folders when the space is depleted. This option is by default disabled.
This commit should solve issue #317.

Co-authored-by: Urmas Rist <urmas@leil.io>

Signed-off-by: Crash <crash@leil.io>