refactor: Make read delayed ops more granular
The read_data_delayed_ops iterates through all the ReadRecord (files
being read currently) and makes some maintenance on them:
- removes the ones ready to be removed
- clears the cache whenever possible
and it was always sleeping for 1/3 of a second after it.

The proposal is to make the work almost always exactly every 1/6 of a
second, but sleeping at least 50ms between every two runs. The previous
version was taking (1/3s + time taken by the actual cycle itself)
between a run of the cycle and the next one.

Notice that the kMaxThresholdTicks was doubled to reflect the higher
frequency of the read delayed ops cycle.

Signed-off-by: Dave <dave@leil.io>