fix: Make ReadCache::Entry::timer atomic
This commit targets making some operations not dependant on other ones
that are mostly unrelated. For instance, to check whether an entry is
ready to be trashed we need to check if it is expired or not and this
was waiting until the request related to this entry is done, which
could to an unexpected overhead.

Signed-off-by: Dave <dave@leil.io>