fix: Add check for safer path by inode on master
Previous version of `fs_full_path_by_inode` was not checking the case
when an inode was already deleted from the system but its current path
by inode value was requested. This could lead to possible dereference
of null pointer, provoking a coredump on the master server. This commit
prevents that issue.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>