test: Fix flaky metadata dumping test (#412)
Previously, test_metadata_prevent_dump_and_chunks_registration failed
randomly because one step was sensitive to the wait for the next
metadata dump cycle, which for some runs was too small, causing the
test to fail.

This commit introduces an extra wait for the next metadata dump cycle
to make sure there is enough time to produce the expected log during
the metadata dump.

Signed-off-by: Crash <crash@leil.io>