fix: Fix data race when ending writes
The release of the inodeLock in the write_data_end could cause some
data race conditions when several file descriptors point to the same
inode. The release of the inodeLock (using the inodeData's mutex) was
being performed after the inodeData is freed. The solution is to
substitute the code in order to avoid it.

Signed-off-by: Dave <dave@leil.io>