feat(mount): New chunk-based write algorithm (#158)
This PR targets the following issues:
- low performance in single thread writes.
- [Windows] stop/freeze of writes when multiple copies are performed at
the same time.
- [Windows] decrease of performance sometimes when the number of
writing threads increases.
It also adds a new parameter for tuning the write wave timeout
(sfschunkserverwavewriteto). Another one, for chosing which write
algorithm to use, was also added (sfsuseoldwritealgorithm).

The changes performed are the following:
- create ChunkData struct to represent and gather the data of the
pending write operation on a chunk. This struct contains many fields
from the inodedata, since it is taking over such responsabilities. The
use of this new structure is the main reason of the wide changes in the
writedata.cc file.
- change of the approach in the write jobs queue: each entry, instead
of being an inodedata one, it is a ChunkData one.
- minor changes in other classes (WriteChunkLocator and ChunkWriter).

A test on very small parallel random writings was added. I've also
checked that there are no new data races conditions added by this
changes.

Signed-off-by: Dave <dave@leil.io>