fix(uRaft): Prevent electors from starting MDS
Previously, the `saunafs-uraft` service started `saunafs-ha-master`
unconditionally, without respecting the `URAFT_ELECTOR_MODE` setting.
As a result, the master service was incorrectly started on elector
nodes, which should only participate in leader elections and do not
start the master process.

This behavior required either manual intervention to stop the master
on electors or custom modifications to the `saunafs-uraft` systemd unit
to decouple it from `saunafs-ha-master`.

This commit updates the `saunafs-uraft-helper` script to manage the
master service based on the `URAFT_ELECTOR_MODE` setting defined in the
uRaft configuration file (`saunafs-uraft.cfg`).

To apply these changes effectively, the systemd unit of the
`saunafs-ha-master` service must be updated to use the new
functions defined in the uRaft helper script.

Co-authored-by: Antuan <antuan@leil.io>
Co-authored-by: Baldor <baldor@leil.io>
Co-authored-by: guillex <guillex@leil.io>

Signed-off-by: Crash <crash@leil.io>