fix(restore.cc): Correctly convert int to uint8_t in changelog restore
When parsing changelogs, the do_<operation> returns int's. Sometimes
these are within range of SAUNAFS_ERROR_CODES (which is max 54). But
sometimes these return -1 when they have trouble parsing the lines (or
potentially other unknown values). saunafs_error_string however only
accepts uint8_t, which means there's an implicit conversion and could
cause unknown/faulty error messages.

This fixes that by explicitly checking the -1 to mean failure in parsing
lines. For other values, an assert is used when NDEBUG is set (so the
programmer is aware that they are returning a wrong value) or the
reasonable default of UNKNOWN error with a log explaining why.

Signed-off-by: Urmas Rist <urmas@leil.io>