test(tools): Fix saunafs command tests for Windows (#364)
Previous changes for tests for the saunafs command on for Windows were
having a flaky behavior for changes from commit bd88d8b. Due to that,
the changes from that commit were reverted in commit 8f2b674.
This commit reintroduces the corresponding changes so that a first set
of the stable tests can be added of now. The idea is to add tests by
groups of them, wich will be faster approach than trying to directly
integrating all tests at once.

Signed-off-by: rolysr <rolysr@leil.io>