feat: Enhance statfs cache handling
Previous changes to the statfs cache did not account for
the group ID and inode when requested. This could lead to
issues, such as the statfs call returning the same output
for a user requesting quotas on the same inode, even when
different values should be displayed. This commit fixes
that issue.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>