feat: Update migration script and add documentation (#428)
Previous script for downgrading was specific to our official Ubuntu
packages, nor did it account for metalogger. The logic of apt
installations and service management will be moved to the dedicated
debian package repository and updated to work with the new script, while
this new script should be distro agnostic in handling the changelogs.

* Add a lot more error checking to the script, including making sure
  master and metalogger are not running.
* Backups will be backed up by default to /var/lib/ instead of /tmp/,
  which is temporary and can be lost if machine is rebooted.
* Add technical documentation regarding migration

Signed-off-by: Urmas Rist <urmas@leil.io>