fix(master): Correct default listen host and port
Due to changes introduced in deprecation commit a5ed228, it became
necessary to handle the scenario where the configuration for the
master host and port to listen for connections are not defined in
its default configuration file. This commit addresses that scenario,
ensuring the master server correctly handles cases where those values
are not explicitly defined in the configuration file.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>