test: Simplify glibc malloc arenas client test (#440)
The previous version of `test_limit_glibc_malloc_arenas_client` test
from the SingleMachineTests suite included unnecessary complexity
when verifying the behavior of the `limitglibcmallocarenas` option.
This complexity made it difficult to assess the test's stability and
flakiness during repeated executions in the CI environment.

This commit simplifies the test by streamlining the scenario while
maintaining the same functionality and ensuring clearer results.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>