feat(chunkserver): Implement listeners for JobPool (#444)
This change targets changing the current behavior of the JobPool class,
in the way that it can serve different threads at the same time.
Current implementation does not allow a single JobPool instance to
serve different threads because it would mix up the wake up calls and
it would cause dangerous inconsistencies.

The proposed implementation partitions the JobPool inner structures for
each of the sources of jobs (which could be master or client network
workers) and creates specific channels of comunication to ensure that
only the source of a job is awaken.

The multiplicity of sources of jobs of a single JobPool needs to be
tested, and this commit enchances their unittests to check it.

Signed-off-by: Dave <dave@leil.io>