ci: Preserve core dumps on Ganesha crashes (#441)
The Ganesha suite sometimes trigger segfaults in CI, but Github Actions
discards core dumps by default, preventing root-cause analysis.

This commit enables the preservation of core dumps generated in case a
segfault occurs, allowing to generate and save the backtrace of the
core dump using gdb.

Additionally, the pipeline prints part of the backtrace file to the
workflow logs, making it easier to quickly inspect the most relevant
part of the backtrace.

Signed-off-by: Crash <crash@leil.io>