fix: Fix master reload crash (#450)
Due to changes introduced in deprecation commit a5ed228, it became
necessary to handle the scenario where the configuration for the
master host and port to listen for connections are not defined in
its default configuration file when reloading the configuration. This
commit addresses that scenario, which could cause a master crash when
releasing the ListenPort or ListenHost variables a few lines below.

Signed-off-by: Dave <dave@leil.io>