feat(mds,chunkserver): Add CLUSTER_ID config var
This change adds the CLUSTER_ID configuration variable for metadata
(master, shadow) and chunkservers. The new variable should be set at
the beginning and it is not intended to be changed frequently.

Metadata servers only accept connections from chunkservers with the
same CLUSTER_ID (sent during registration), this prevents connections
from undesired chunkservers.

The variable defaults to `default`, to not break compatibility with
existing installations.

Signed-off-by: guillex <guillex@leil.io>