feat(cgi)!: Deprecate cgi for LizardFS v3.12.0
This commit removes all code on the cgi monitoring module related
with the LizardFS versions equal or older than v3.12.0. This
change will be part of the SaunaFS v5.0.0 release, where support
to older LizardFS versions is not longer required.

BREAKING CHANGE: Master servers compiled with LizardFS v3.12.0
or older may not be able to work with this monitoring version.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>