chore: Add alias for `saunafsVersion(5,0,0)`
This commit replaces occurrences of the hardcoded
`saunafsVersion(5,0,0)` in the code with a descriptive
constant, specifically for the introduction of displaying
the content of `.saunafs_mount_info` on the monitoring
service for each mountpoint. Also, this commit includes a
similar change on the CGI monitoring module regarding
naming this hardcoded version.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>