feat(mds): Separate changelogs from metadata backend
Some of the functions related to changelogs were not correctly declared
in IMetadataBackend and implemented in MetadataBackendFile.

This change moves those functions to the proper place, which improves
the modularity and simplifies the interface for metadata backends.

The code was just moved with minimal changes to ease the review. The
logic remains the same.

Signed-off-by: guillex <guillex@leil.io>