fix(mds): Add init function to metadata backends (#475)
This commit reduces the concrete code depending on specific metadata
file based backend, like references to `metadata.sfs`, and delegates
it to the concrete backend implementation, by using the new `init`
member function.

This way, the dependencies on specific filenames for the metadata,
outside of the concrete backend implementations, are reduced.
Implementing and integrating new backends should be easier now.

Signed-off-by: guillex <guillex@leil.io>