feat: Implement InputBuffer for multi block write
This commit implements the InputBuffer class that would be useful for
making the chunkserver writing several blocks at a time. Main idea for
it is to have the raw block data in consecutive and aligned memory. The
rest of the structure is there to help the process of enqueuing the
related write job and protecting the update of the InputBuffer.

The InputBuffer is supposed to be fed by data coming directly from the
client (by reading from the socket) which is a network worker thread
responsability, and be consumed by a JobPool worker thread.

This commit also renames output_buffer* files to better describe the
structures implemented inside them. The includes to old output_buffer.h
were also updated.

Signed-off-by: Dave <dave@leil.io>