feat: Use InputBuffer in writes
This commit rewrites part of the chunkserver write algorithm to use the
InputBuffer class. The main goal is to prepare it to load a single
InputBuffer with several blocks of data that are expected to be written
and then make the JobPool workers able to write as much data at once as
possible.

Signed-off-by: Dave <dave@leil.io>