fix(chunkserver): Fallback to old registration if needed (#489)
Fixes new Chunkservers connecting to Master in versions 4.x.

This change modifies the registration process in the chunkservers to
fallback to the old procedure if the Master is not answering to the
registration request. A possible cause is that Master runs a version
before 5.x.

The test was modified to cover the initial old Master rejection, the
succesfull Chunkserver reconnection using the old procedure and a
reconnection after Master restarts.

Signed-off-by: guillex <guillex@leil.io>