fix(mount)!: Redirect info messages to stdout
Previously, some info messages during registration process of the client
were being sent to stderr, which is not accurate given that these
messages are not errors but rather informational. This commit fixes that
issue.

BREAKING CHANGE: Some older tools that expect to catch client output
related to master registration from stderr may need to be modified to
get this information from stdout.

Signed-off-by: rolysr <rolysr@leil.io>