fix: Fix write algorithm in delayed init mode
This commit targets fixing the inconsistent selection of write
algorithm when using the sfsdelayedinit option in client.

Main feature of sfsdelayedinit option is to be able to start the mount
without an available master, which means it is not expected after short
notice to know the master version. This means client is not supposed to
display any warning message right after mounting and select the
appropiate write algorithm after knowing the current master version.

This commit also implements the support for the client to switch the
write algorithm used when needed depending on master version.

Signed-off-by: Dave <dave@leil.io>