chore(master): Improve logging for oversized packet
This commit improves the log message in the master server by including
the peer's IP address and the packet type when an oversized packet is
received. This makes debugging easier by identifying the client source.

It also introduces a helper function to convert the IPv4 address from
matoclserventry::peerIpAddress to a readable string.

Signed-off-by: Crash <crash@leil.io>