fix: Fix write algorithm in delayed init mode (#496)
This PR targets fixing the inconsistent selection of write
algorithm when using the `sfsdelayedinit` option in client.

Main feature of sfsdelayedinit option is to be able to start the mount
without an available master, which means it is not expected after short
notice to know the master version. This means client is not supposed to
display any warning message right after mounting and select the
appropiate write algorithm after knowing the current master version.

Implements the support for the client to switch the
write algorithm used when needed depending on master version.

An integration was added and another one was enhanced in order
to test the proposed changes. Also, a specific test for the
`sfsdelayedinit` option was added in order to check its expected
behavior.

Signed-off-by: Dave <dave@leil.io>