feat(common): Add initial signal-slots support (#507)
Adds an initial implementation of observable properties using a
signal-slot mechanism. With this approach, the observed object
doesn’t need to know its observers. For example, the metadata
version in gMetadata could be an ObservableProperty, and at
runtime different observers would be notified and react to
changes.

This will be used by the FoundationDB metadata backend to persist
changes immediately in the database instead of relying on the
current dump mechanism.

Co-authored-by: Crash <crash@leil.io>

Signed-off-by: guillex <guillex@leil.io>