test: Fix path by inode test on Windows
Due to client unmount logic introduced in the path-by-inode related test
in commit ebc58a61, this test was failing in the Windows client test
suite. On WSL1, the return value from assert_success was not handled
correctly when the client was restarted. Additionally, asserting that
specific call is unnecessary, as the test will fail if the client is
not mounted or does not include the required option for mounting to the
given subfolder. This commit resolves these issues for the Windows
client test suite.

Signed-off-by: rolysr <rolysr@leil.io>