fix: Fix upgrade master switch write algo
It was noticed that master was not restarting in some failed runs of
the test test_upgrade_master_switch_write_algo. This commit makes sure
that test uses the usual saunafs_master_daemon.

The main change was to avoid calling again the is_windows_system
function, so it was stored at the begining of the
setup_local_empty_saunafs. This way many tests were changed the way it
used the new stored value.

Signed-off-by: Dave <dave@leil.io>